jupyter

+++ deploy.sh
@@ -1,5 +1,7 @@
-curl -L -o jbake.zip https://github.com/jbake-org/jbake/releases/download/v2.6.7/jbake-2.6.7-bin.zip
-unzip jbake.zip && rm -f jbake.zip
+if [ ! -d jbake-2.7.0-rc.7-bin ]; then
+	curl -L -o jbake.zip https://github.com/jbake-org/jbake/releases/download/v2.7.0-rc.7/jbake-2.7.0-rc.7-bin.zip
+	unzip jbake.zip && rm -f jbake.zip
+fi
 
 jbake-*/bin/jbake -b
